Cleanup
Moved ActiveTheme to ViewModel. This makes it easier to find via the data context instead of the control hierarchy, and I think it's appropriate.
Separated MediaDisplay from MediaControls in preparation for the playlist view which only has the latter.
Moved all the ad-hoc styles that fetch theme from different places to unified ones in Styles.xaml.
Added SizeHost control where buttons inherit their sizes from, instead of from the MediaDisplay, since not all buttons are part of it.