Finally seems to work properly
Long difficult debugging session.
PlaceholderLoadable now shows placeholder on failure, so binding doesn't freak out.
MutableStream is no longer closed directly, it's closed in SongFile and replaced with a new loadable.
BeginInvoke seems to work as well as Invoke, and order is preserved.
CurrentTrack is now cached because the playlist shuffling around made the index invalid for brief moments.
Re-added resampling, it's not as nice as before but should be consistent.